pyproject.toml
[project]
name = "ninuka-quiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "spacy>=3.7.0",
    "umap-learn>=0.5.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "numpy>=1.26.0",
    "nltk>=3.8.0",
    "wordcloud>=1.9.0",
]
